- Add the control of the verbosity of the sdialog.audio module by creating its own logger - Restore the verbosity level after the call of the "to_audio" and "AudioPipeline.inference()" methods - Rename the module "tts_engine" into "tts" and make it fetching the TTS classes in a directory "tts" in order to simplify collaboration in the future - Clean the import order for all the audios files - Fix documentation after renaming "tts_engine" into "tts" - Fix documentation for the TTS sub-directories - Update and run all the tests locally
