+++ src/sdialog/audio/jsalt.py
@@ -49,6 +49,7 @@ import math
 import random
 from enum import Enum
 from typing import Tuple, Dict, Any, Optional
+
 from sdialog.audio.room import Room, Dimensions3D
 from sdialog.audio.room_generator import RoomGenerator
 from sdialog.audio.utils import Furniture, RGBAColor, RoomMaterials
