+++ src/sdialog/audio/room_generator.py
@@ -33,6 +33,7 @@ import time
 import random
 from abc import abstractmethod
 from typing import Tuple, Dict, Any, Optional
+
 from sdialog.audio.utils import Furniture, RoomMaterials
 from sdialog.audio.room import Room, Dimensions3D, MicrophonePosition
