+++ src/sdialog/audio/tts/__init__.py
@@ -1,0 +1,6 @@
+from .base import BaseTTS
+from .huggingface import HuggingFaceTTS
+from .index import IndexTTS
+from .kokoro import KokoroTTS
+
+__all__ = ["BaseTTS", "KokoroTTS", "IndexTTS", "HuggingFaceTTS"]
