+++ src/sdialog/audio/tts/huggingface/__init__.py
@@ -1,0 +1,3 @@
+from .tts import HuggingFaceTTS
+
+__all__ = ["HuggingFaceTTS"]
