+++ src/sdialog/audio/tts/index/__init__.py
@@ -1,0 +1,3 @@
+from .tts import IndexTTS
+
+__all__ = ["IndexTTS"]
