+++ src/sdialog/audio/tts/kokoro/__init__.py
@@ -1,0 +1,3 @@
+from .tts import KokoroTTS
+
+__all__ = ["KokoroTTS"]
